+++ test/eslint/rules/jsdoc-type-application-dot.js
@@ -1,0 +1,80 @@
+const {errors, concat} = require(`../helper`);
+
+module.exports = {
+	valid: [{
+		options: ['always'],
+		code: concat(
+			`/**`,
+			` * @return {Array.<string>}`,
+			` */`
+		)
+	}, {
+		options: ['never'],
+		code: concat(
+			`/**`,
+			` * @return {Array<string>}`,
+			` */`
+		)
+	}],
+	invalid: [{
+		options: ['always'],
+		code: concat(
+			`/**`,
+			` * @return {Array<string>}`,
+			` */`
+		),
+		output: concat(
+			`/**`,
+			` * @return {Array.<string>}`,
+			` */`
+		),
+		errors: errors(
+			`Type application should have dot before "<"".`
+		)
+	}, {
+		options: ['always'],
+		code: concat(
+			`/**`,
+			` * @return {Array. <string>}`,
+			` */`
+		),
+		output: concat(
+			`/**`,
+			` * @return {Array. .<string>}`,
+			` */`
+		),
+		errors: errors(
+			`Type application should have dot before "<"".`
+		)
+	}, {
+		options: ['never'],
+		code: concat(
+			`/**`,
+			` * @return {Array.<string>}`,
+			` */`
+		),
+		output: concat(
+			`/**`,
+			` * @return {Array<string>}`,
+			` */`
+		),
+		errors: errors(
+			`Type application should not have dot before "<"".`
+		)
+	}, {
+		options: ['never'],
+		code: concat(
+			`/**`,
+			` * @return {Array .<string>}`,
+			` */`
+		),
+		output: concat(
+			`/**`,
+			` * @return {Array <string>}`,
+			` */`
+		),
+		errors: errors(
+			`Type application should not have dot before "<"".`
+		)
+	}]
+};